Move non private extensions into new files
